requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
opencv-python==4.8.1.78
numpy==1.26.2
torch==2.2.0
Pillow==10.1.0
transformers==4.35.0 